+++ apps/web/src/app/(dashboard)/layout.tsx
@@ -5,6 +5,8 @@ import { redirect } from "next/navigation"
 import { Header } from "@/components/dashboard/Header"
 import { Sidebar } from "@/components/dashboard/Sidebar"
 import { coreRoutes } from "@/global/routes"
+import { DashboardStateProvider } from "./controller"
+import { resolveSignedInUserDetails } from "./serverActions"
 
 
 export default async function RootLayout({
@@ -13,27 +15,41 @@ export default async function RootLayout({
   // Check if we are signed in
   const signedIn = await auth.signedIn()
   if (!signedIn) redirect('/sign-in')
+  const { user, organization } = await resolveSignedInUserDetails()
+
+  console.log(`organization.data`, organization.data)
 
   // Create a sidebar for the mobile view
   const compactSidebar = (
-    <Sidebar 
-      className="w-full" 
-      links={coreRoutes} 
+    <Sidebar
+      className="w-full"
+      links={coreRoutes}
     />
   )
   // Otherwise serve the route
   return (
-    <div className="grid grid-cols-[0_1fr] md:grid-cols-[auto_1fr] w-full h-full min-h-[100vh]">
-      <Sidebar compacts 
-        className="hidden md:flex" 
-        links={coreRoutes} 
-      />
-      <div className="w-[100vw] md:w-full h-[100vh] grid grid-rows-[auto_1fr]">
-        <Header hasMenuButton sidebar={compactSidebar} />
-        <div className="w-full h-full overflow-y-auto">
-          {children}
+    <DashboardStateProvider initial={{
+      organizationId: organization.id() as string,
+      organizationHandle: organization.data.handle || '',
+      googleLink: organization.data.googleLink || '', 
+      instagramHandle: organization.data.instagramHandle || '',
+      tikTokHandle: organization.data.tikTokHandle || '',
+      organizationUsers: [],
+      history: [],
+      campaigns: [],
+    }}>
+      <div className="grid grid-cols-[0_1fr] md:grid-cols-[auto_1fr] w-full h-full min-h-[100vh]">
+        <Sidebar compacts
+          className="hidden md:flex"
+          links={coreRoutes}
+        />
+        <div className="w-[100vw] md:w-full h-[100vh] grid grid-rows-[auto_1fr]">
+          <Header hasMenuButton sidebar={compactSidebar} />
+          <div className="w-full h-full overflow-y-auto">
+            {children}
+          </div>
         </div>
       </div>
-    </div>
+    </DashboardStateProvider>
   )
 }