+++ apps/web/src/models/Organization.ts
@@ -3,6 +3,8 @@
 import { baseModel } from "@repo/models";
 import { z } from "zod";
 import { dynamodbConnection } from '@/config'
+import { User } from "./User";
+import { shortId } from "@repo/utilities";
 
 
 export const OrganizationSchema = z.object({
@@ -22,7 +24,7 @@ export const OrganizationSchema = z.object({
 export type OrganizationType = z.infer<typeof OrganizationSchema>
 
 
-export class Organization extends baseModel({
+export class Organization extends baseModel<OrganizationType>({
   name: 'Organization',
   schema: OrganizationSchema,
   connection: dynamodbConnection,
@@ -30,4 +32,33 @@ export class Organization extends baseModel({
     { partition: 'id' },
     { partition: 'handle', sort: 'dateCreated' },
   ],
-}) { }
+}) { 
+
+  async users () {
+    // Fetch the users from this organization 
+    return this
+  }
+
+  static async getUserOrganization (user: User) {
+    // Make sure the user has an organization assigned 
+    const organizationId = user.data.organizationId || shortId()
+    const userHasNoOrganization = user.data.organizationId == null
+    if (userHasNoOrganization) {
+      user.data.organizationId = organizationId
+      await user.push()
+    }
+
+    // Make sure the organization exists
+    const organization = new Organization({ 
+      id: organizationId,
+      googleLink: '',
+      instagramHandle: '',
+      tikTokHandle: '',
+      handle: shortId(),
+    })
+    const exists = await organization.exists()
+    if (!exists) await organization.create()
+    await organization.pull()
+    return organization
+  }
+}