+++ packages/database/test/index.ts
@@ -1,0 +1,4 @@
+
+export * from './setup'
+import dynamoSchema from './dynamo.schema.json'
+export { dynamoSchema }