+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Made the customer view page visuals
 - Made a data table component
 ### Changed
-- The application is now a PWA
+- [upstream] The application is now a PWA
 ### Fixed
+- Added type safety to the data models "data" field by inferring the data type
+- [upstream] tests for @repo/database and @repo/models now work
+- [upstream] model.exists now tests all GSI's to avoid model duplicates